debian: wheezy: Needed to seperate soem of the depends due to package naming differences
wheezy and jessie required different pacakge names for some packages.  To be
able to keep all the depends in the contol file and not have to have them applied
within prepare-chroot-debian I added ann X-tra variable field for paragraph 1 in
the Debian control file which describes the source

These field is 'X-Qubes-Build-Depends-<distro>:  where distro is the name of the
distribution such as 'wheezy' or 'jessie'.  Multiple distros can be listed naming
their specific depends.

A function within the Makefile.debian/qubuntu will apply the depends.